DF-9998 sync win32 and macos implementations
Former-commit-id: 6376339317d80c9ecbe91fcbefe9ea1e44a79c72

+++ Sources/Internal/Engine/Private/Win32/EventWin32.h
@@ -25,8 +25,8 @@ struct EventWin32 final
 
     struct ResizeEvent
     {
-        int32 width;
-        int32 height;
+        float32 width;
+        float32 height;
     };
 
     eType type = DUMMY;
